move Notify services/adapters into services
