+++ apps/pushnotif-func/src/main.ts
@@ -41,17 +41,17 @@ import { handle as handleNotifyMessage } from "./functions/handle-nh-notify-mess
 import { Info } from "./functions/info";
 import { Notify } from "./functions/notify";
 import { createClient } from "./generated/session-manager/client";
-import { initTelemetryClient } from "./utils/appinsights";
-import { getConfigOrThrow } from "./utils/config";
-import { cosmosdbClient, cosmosdbInstance } from "./utils/cosmosdb";
-import { NotificationHubPartitionFactory } from "./utils/notificationhub-service-partition";
-import { sendNotification } from "./utils/notify/notification";
+import { sendNotification } from "./service/notification";
 import {
   getMessageWithContent,
   getService,
   getUserProfileReader,
   getUserSessionStatusReader,
-} from "./utils/notify/readers";
+} from "./service/readers";
+import { initTelemetryClient } from "./utils/appinsights";
+import { getConfigOrThrow } from "./utils/config";
+import { cosmosdbClient, cosmosdbInstance } from "./utils/cosmosdb";
+import { NotificationHubPartitionFactory } from "./utils/notificationhub-service-partition";
 
 // ---------------------------------------------------------------------------
 // Shared configuration and Functions dependencies
